update to social history and privacy level

+++ app/routers/social_history_sensitive_mapping_router.py
@@ -9,27 +9,32 @@ from ..auth.jwt_utils import extract_jwt_payload, get_user_id, get_role_name
 
 router = APIRouter()
 
+@router.get("/social_history_sensitive_mapping", response_model=list[SocialHistorySensitive])
+def read_all_social_history_sensitive_mapping(db: Session = Depends(get_db)):
+    db_social_history_sensitive_mapping = get_all_social_history(db)
+    if db_social_history_sensitive_mapping is None:
+        raise HTTPException(status_code=404, detail="Social History Mapping not found")
+    return db_social_history_sensitive_mapping
+
 @router.post("/social_history_sensitive_mapping/add", response_model=SocialHistorySensitive)
-def create_new_social_history_sensitive_mapping(request: Request, social_history_item: str, social_history_sensitive_mapping: SocialHistorySensitiveCreate, db: Session = Depends(get_db), require_auth: bool = True):
-    # payload = extract_jwt_payload(request, require_auth)
-    # user_id = get_user_id(payload)
-    # role_name = get_role_name(payload)
+def create_new_social_history_sensitive_mapping(request: Request, social_history_sensitive_mapping: SocialHistorySensitiveCreate, db: Session = Depends(get_db), require_auth: bool = True):
+    payload = extract_jwt_payload(request, require_auth)
+    role_name = get_role_name(payload)
     
-    # is_admin = role_name == "ADMIN"
-    # if not is_admin:
-    #     raise HTTPException(status_code=404, detail="User is not authorised")
+    is_admin = role_name == "ADMIN"
+    if not is_admin:
+        raise HTTPException(status_code=404, detail="User is not authorised")
     
-    return create_sensitive_mapping(db=db, social_history_item=social_history_item, social_history=social_history_sensitive_mapping, created_by=1)
+    return create_sensitive_mapping(db=db, social_history=social_history_sensitive_mapping, created_by=1)
 
 @router.put("/social_history_sensitive_mapping/update", response_model=List[SocialHistorySensitive])
 def update_all_social_history_sensitive_mapping(request: Request, social_history_sensitive_mapping_list: List[SocialHistorySensitiveUpdate], db: Session = Depends(get_db), require_auth: bool = True):
-    # payload = extract_jwt_payload(request, require_auth)
-    # user_id = get_user_id(payload)
-    # role_name = get_role_name(payload)
+    payload = extract_jwt_payload(request, require_auth)
+    role_name = get_role_name(payload)
     
-    # is_admin = role_name == "ADMIN"
-    # if not is_admin:
-    #     raise HTTPException(status_code=404, detail="User is not authorised")
+    is_admin = role_name == "ADMIN"
+    if not is_admin:
+        raise HTTPException(status_code=404, detail="User is not authorised")
     
     db_social_history_sensitive_mapping = update_all_sensitive_mapping(db=db, social_history_list=social_history_sensitive_mapping_list, modified_by=1)
     if db_social_history_sensitive_mapping == []:
@@ -39,13 +44,12 @@ def update_all_social_history_sensitive_mapping(request: Request, social_history
 
 @router.delete("/social_history_sensitive_mapping/delete/{social_history_item}", response_model=SocialHistorySensitive)
 def delete_existing_social_history_sensitive_mapping(request: Request, social_history_item: str, db: Session = Depends(get_db), require_auth: bool = True):
-    # payload = extract_jwt_payload(request, require_auth)
-    # user_id = get_user_id(payload)
-    # role_name = get_role_name(payload)
+    payload = extract_jwt_payload(request, require_auth)
+    role_name = get_role_name(payload)
     
-    # is_admin = role_name == "ADMIN"
-    # if not is_admin:
-    #     raise HTTPException(status_code=404, detail="User is not authorised")
+    is_admin = role_name == "ADMIN"
+    if not is_admin:
+        raise HTTPException(status_code=404, detail="User is not authorised")
     
     db_social_history_sensitive_mapping = delete_sensitive_mapping(db=db, social_history_item=social_history_item)
     if db_social_history_sensitive_mapping is None:
